feat(client): add styled
